Change folder structure: Move user selection, malltalk setup and settings and step logic to their own folders
