+++ frontend/src/components/malltalk/interfaces/MyTableSettings.ts
