+++ frontend/src/components/malltalk/settings/ChangeUsers.vue
@@ -10,12 +10,12 @@
 
 <script lang="ts" setup>
 import SimpleButton from '#components/buttons/SimpleButton.vue'
+import UserSelection from '#components/user-selection/UserSelection.vue'
 import GlobalErrorHandler from '#plugins/globalErrorHandler'
-import UserSelection from '#src/panels/components/UserSelection.vue'
 import { useTablesStore } from '#stores/tablesStore'
 
-import type { StepProps, StepEmits } from '#src/panels/composables/useSteps'
-import type MyTableSettings from './interfaces/MyTableSettings'
+import type MyTableSettings from '#components/malltalk/interfaces/MyTableSettings'
+import type { StepEmits, StepProps } from '#components/steps/StepComponentTypes'
 
 const props = defineProps<StepProps>()
 const emit = defineEmits<StepEmits>()
