+++ frontend/src/components/malltalk/setup/TableSetup.vue
@@ -1,5 +1,5 @@
 <template>
-  <PanelHeader
+  <StepHeader
     v-if="steps"
     :title="steps[currentStep]?.title ?? 'unknown'"
     :is-back-button-visible="currentStep > 0"
@@ -21,15 +21,17 @@
 import { navigate } from 'vike/client/router'
 import { reactive } from 'vue'
 
+import StepHeader from '#components/steps/StepHeader.vue'
+import { Step, useSteps } from '#components/steps/useSteps'
 import GlobalErrorHandler from '#plugins/globalErrorHandler'
-import { Step, useSteps } from '#src/panels/composables/useSteps'
-import MyTableSettings from '#src/panels/dreammall/interfaces/MyTableSettings'
-import TableSetupStepA from '#src/panels/dreammall/TableSetupStepA.vue'
-import TableSetupStepB from '#src/panels/dreammall/TableSetupStepB.vue'
-import TableSetupStepC from '#src/panels/dreammall/TableSetupStepC.vue'
-import TableSetupStepD from '#src/panels/dreammall/TableSetupStepD.vue'
-import PanelHeader from '#src/panels/PanelHeader.vue'
 import { useTablesStore } from '#stores/tablesStore'
+
+import TableSetupStepA from './TableSetupStepA.vue'
+import TableSetupStepB from './TableSetupStepB.vue'
+import TableSetupStepC from './TableSetupStepC.vue'
+import TableSetupStepD from './TableSetupStepD.vue'
+
+import type MyTableSettings from '#components/malltalk/interfaces/MyTableSettings'
 
 const tablesStore = useTablesStore()
 
@@ -112,3 +114,4 @@ const onClose = () => emit('close')
 reset()
 defineExpose({ reset })
 </script>
+#src/panels/dreammall-button-drawer/interfaces/MyTableSettings.js
