+++ frontend/src/components/malltalk/setup/TableSetupStepA.vue
@@ -10,7 +10,7 @@
 <script setup lang="ts">
 import SimpleButton from '#components/buttons/SimpleButton.vue'
 
-import type { StepEmits } from '#src/panels/composables/useSteps'
+import type { StepEmits } from '#components/steps/StepComponentTypes'
 
 const emit = defineEmits<StepEmits>()
