+++ frontend/src/components/malltalk/setup/TableSetupStepB.vue
@@ -21,8 +21,8 @@
 <script setup lang="ts">
 import SimpleButton from '#components/buttons/SimpleButton.vue'
 
-import type { StepProps, StepEmits } from '#src/panels/composables/useSteps'
-import type MyTableSettings from '#src/panels/dreammall/interfaces/MyTableSettings'
+import type MyTableSettings from '#components/malltalk/interfaces/MyTableSettings'
+import type { StepEmits, StepProps } from '#components/steps/StepComponentTypes'
 
 defineProps<StepProps>()
 const emit = defineEmits<StepEmits>()
@@ -39,4 +39,4 @@ const onNext = () => {
   box-shadow: none !important;
 }
 </style>
-./StepProps
+./StepProps #src/panels/dreammall-button-drawer/interfaces/MyTableSettings.js
