+++ frontend/src/components/steps/StepComponentTypes.ts
@@ -1,0 +1,8 @@
+export type StepProps = {
+  submitText: string
+}
+
+export type StepEmits = {
+  (e: 'next' | 'submit'): void
+  (e: 'goTo', value: string): void
+}
