+++ frontend/src/components/steps/StepHeader.vue
